Add support for alias
With the way the steps are validated, the alias can never be used as a
step which avoided a bug.

Also, instead of doing a replace of the alias to number like in other
libs, we simply validate the alias as the last step before validating
the number. This way we avoid adding behavoir for places where the alias
would be of value 1 in a preset and value 0 in another (e.g.: sun = 1 in
aws, sun = 0 everywhere else).

Some libraries can accept the full names (such as node-cron), however
they also accept the 3 char names, so for simplicity sake, we only
support short alias for now.

If we ever want to support variable aliases, we could use the preset
configurations to allow overriding the aliases list.