feat: load presets on module load
In the previous source code version we would load the built-in presets on every
invocation of 'cron' (via 'validateOptions'). This means that we do a lot of
repeated work on consecutive invocations of 'cron'. With the new design we load
the presets just once when we load the module 'options.ts', via the new function
'loadPresets()'.

In order to avoid circular dependencies, I modified 'presets.ts' to export
just vanilla JavaScript objects.

Signed-off-by: albertodiazdorado <albertodiazdorado@gmail.com>