feat: improve performance by adding a cache mechanism
cron-validated was very slow when called in a loop. The reason is that we
repeated all these operations in every call:
1. Merge the preset and the overrides into a so-called 'unvalidatedConfig'
2. Dynamically generate a yup validation schema from the preset and the override
3. Validate the 'unvalidatedConfig' using the dynamically generated schema

Step 2 is particularly slow, but steps 1 and 3 aren't particularly fast either.
Luckily for us, all these 3 steps are perfectly cacheable. This Pull Request
introduces a so-called 'optionsCache' to avoid repeating the 3 steps if we
already executed them in the past (which happens all the time in a loop).

The cache is a map from 'cache keys' to 'validated options'. The cache key
is the concatenarion of 'presetId' and a stringified 'override' object. If two
function calls produce the same cache key, it invariably means that the user
wants to use the same 'Options' in both function calls. This is how the cache
produces a performance improvement of about x100 for consecutive calls.

Note that the construction of the cache key is not optimized. If the user uses
'cron-validate' two consecutive times with identical 'override' objects, but
swaps the order of the override object fields, then they will not benefit from
caching. This can be fixed by using deterministic stringification, but I do not
consider it necessary.

Signed-off-by: albertodiazdorado <albertodiazdorado@gmail.com>